# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forum API",
        "description": "API for the forum application",
        "version": "1.0.0"
    },
    "paths": {
        "/comments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Store a newly created comment in storage.",
                "operationId": "a0c0613c248bc727d51299e2eb03d2e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Some comment"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/posts/create": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Show the form for creating a new resource.",
                "operationId": "87880105c55ee709353642ef648ce4ec",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "c2e0b0e26215e6d34ece33ebae25f7df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Some title"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Some text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Display the specified resource.",
                "operationId": "c8bd68bc05422c879bc2dd429ae5782b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "$ref": "#/components/schemas/Post"
                                        },
                                        "comments": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/edit": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Display profile edit form",
                "description": "Display the user's profile form.",
                "operationId": "d06bd8fd9d326eb7a380edbdcea9928f",
                "responses": {
                    "200": {
                        "description": "Display user's profile form"
                    }
                }
            }
        },
        "/profile": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete user account",
                "description": "Delete the user's account.",
                "operationId": "1d6a5c6489e35a763aa97ebef12863ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "current_password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile information",
                "description": "Update the user's profile information.",
                "operationId": "354f5b23296bed4036564563e7423f43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show the user's profile page",
                "operationId": "d3724f04f04f6ddc1096674d104f8b32",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "ce3cf31668a69eb39e5e98738715afa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Some title"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Some text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/create": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show the form for creating a new resource.",
                "operationId": "8c02c9f610b5c714fa4921bb178f92a6",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display the specified user.",
                "operationId": "1ec60d4ee968175164a94c21f897f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "posts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Comment": {
                "title": "Comment model",
                "description": "Comment model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Some text"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "post_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Comment"
                }
            },
            "Post": {
                "title": "Post model",
                "description": "Post model",
                "properties": {
                    "id": {
                        "description": "The ID of the post",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "The ID of the user who created the post",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "text": {
                        "description": "The text of the post",
                        "type": "string",
                        "example": "Some text"
                    },
                    "title": {
                        "description": "The title of the post",
                        "type": "string",
                        "example": "Some title"
                    },
                    "created_at": {
                        "description": "The date and time the post was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 12:00:00"
                    },
                    "updated_at": {
                        "description": "The date and time the post was updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "Timestamp when the user verified their email",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string"
                    },
                    "remember_token": {
                        "description": "Token to be used for remember me functionality",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the user was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the user was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Comment",
            "description": "Comment"
        },
        {
            "name": "Post",
            "description": "Post"
        },
        {
            "name": "Profile",
            "description": "Profile"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}